User GUI
This Python script implements a simple graphical user interface (GUI) using Tkinter to collect and validate user inputs related to the Octree range. The application provides fields for entering financial and rating ranges, as well as selecting date ranges (in days since 01/01/2017). It also includes an input for maximum capacity and a button to submit the form.